requirements.txt
# pip freeze > requirements.txt

annotated-types==0.7.0
anyio==3.7.1
boto3==1.38.1
botocore==1.38.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.2.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==1.10.13
pydantic_core==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy>=2.0.0
starlette==0.27.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.24.0
strawberry-graphql==0.215.1